Storage/VHD: Fix resize corruption if the image has more than 4GB of blocks allocated
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@52695 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Storage/VHD.cpp
@@ -2696,7 +2696,7 @@ static int vhdResize(void *pBackendData, uint64_t cbSize,
         {
             if (pImage->pBlockAllocationTable[idxBlock] != ~0U)
             {
-                uint64_t offStartBlock = pImage->pBlockAllocationTable[idxBlock] * VHD_SECTOR_SIZE;
+                uint64_t offStartBlock = (uint64_t)pImage->pBlockAllocationTable[idxBlock] * VHD_SECTOR_SIZE;
                 if (offStartBlock < offStartDataOld)
                     offStartDataOld = offStartBlock;
                 cBlocksAllocated++;
